Refactor ReferralStep into ReferralStepForm
