+++ src/api/operations/ce.fragments.graphql
@@ -275,7 +275,9 @@ fragment CeReferralStepFields on CeReferralStep {
   formDefinition {
     ...FormDefinitionFields
   }
-  submittedValues
+  customDataElements {
+    ...CustomDataElementFields
+  }
 }
 
 # used for resolving assigned steps on the User Dashboard
